refactor: フォローボタンをコンポーネント化
